refactor: 상품 상세 페이지 이미지 크기 조정
