+++ .storybook/preview.cjs
@@ -1,4 +1,4 @@
-import '../src/css/index.css';
+import '../src/styles/css/index.css';
 
 export const parameters = {
   actions: { argTypesRegex: '^on[A-Z].*' },
