+++ src/components/atomes/flex/index.tsx
@@ -1,17 +1,17 @@
 import { memo, PropsWithChildren, useCallback } from 'react';
 
 interface IFlex {
-  type?: string;
+  className?: string;
   onClick?: () => void;
 }
 
-function Flex({ children, type = '', onClick }: PropsWithChildren<IFlex>) {
+function Flex({ children, className = '', onClick }: PropsWithChildren<IFlex>) {
   const handleClick = useCallback(() => {
     onClick?.();
   }, []);
 
   return (
-    <div className={`flex ${type}`} onClick={handleClick}>
+    <div className={`flex ${className}`} onClick={handleClick}>
       {children}
     </div>
   );
