+++ src/styles/css/common/index.css
