+++ src/styles/css/common/number-input.css
