+++ src/styles/css/page/product.css
@@ -12,7 +12,8 @@
     display: flex;
     flex-direction: column;
     align-items: center;
-    margin-top: 50px;
+    margin: 50px auto 0;
+    max-width: 500px;
 }
 
 .product-detail-info {
