+++ src/styles/scss/index.scss
@@ -1,0 +1,1 @@
+@import 'card.scss';
